Conversion of HockeyApp to AppCenter for crash reporting
