+++ src/iOS.Core/iOS.Core.csproj
@@ -91,7 +91,7 @@
     <Compile Include="Services\DeviceActionService.cs" />
     <Compile Include="Utilities\ASHelpers.cs" />
     <Compile Include="Utilities\Dialogs.cs" />
-    <Compile Include="Utilities\HockeyAppCrashManagerDelegate.cs" />
+    <Compile Include="Utilities\AppCenterHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\CryptoPrimitiveService.cs" />
     <Compile Include="Services\KeyChainStorageService.cs" />
