+++ src/iOS.Extension/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.AppCenter" version="3.1.0" targetFramework="xamarinios10" />
+  <package id="Microsoft.AppCenter.Crashes" version="3.1.0" targetFramework="xamarinios10" />
+</packages>
