+++ src/iOS/AppDelegate.cs
@@ -276,7 +276,7 @@ namespace Bit.iOS
             RegisterPush();
             var deviceActionService = ServiceContainer.Resolve<IDeviceActionService>("deviceActionService");
             ServiceContainer.Init(deviceActionService.DeviceUserAgent);
-            iOSCoreHelpers.RegisterHockeyApp();
+            iOSCoreHelpers.RegisterAppCenter();
             _pushHandler = new iOSPushNotificationHandler(
                 ServiceContainer.Resolve<IPushNotificationListenerService>("pushNotificationListenerService"));
             _nfcDelegate = new NFCReaderDelegate((success, message) =>
